discord/requirements.txt
gevent
requests
websocket-client